app/requirements.txt
Django==4.1.13
asgiref>=3.5.2
sqlparse==0.5.0
debugpy==1.5.1
gunicorn==22.0.0
mssql-django>=1.4
pyodbc>=4.0
python-dotenv>=1.0
# Third-party integrations for history tracking and enhanced widgets
django-simple-history==3.4.0
django-select2==7.10.0
django-filter==23.5
#certifi==2021.10.8
#cffi==1.15.0
#charset-normalizer==2.0.12
#cryptography==36.0.1
#defusedxml==0.7.1
#django-allauth==0.49.0
#idna==3.3
#oauthlib==3.2.0
#pycparser==2.21
#PyJWT==2.3.0
#python3-openid==3.2.0
#requests==2.27.1
#requests-oauthlib==1.3.1
#urllib3==1.26.8

